Added return (unsigned int)(next / 65536) % ((unsigned int)RAND_MAX + 1u);

+++ kernel/libc/kernel.c
@@ -4,7 +4,7 @@ static unsigned long next = 1;
 
 int rand(void) {
     next = next * 1103515245 + 12345;
-    return (unsigned int)(next / 65536) % (RAND_MAX + 1);
+    return (unsigned int)(next / 65536) % ((unsigned int)RAND_MAX + 1u);
 }
 
 void srand(unsigned int seed) {
